Update Forest_Cover_Classification.ipynb
This model was created as part of codecademy.com "Build Deep Learning Models with Tensorflow" path final portfolio project.
All of the work in here is my own except for the block with the class PlotLearning, I was trying to find a way to plot the loss and metrics during learning as opposed to plotting the 'history' variable afterwards. 
With help from:
https://medium.com/geekculture/how-to-plot-model-loss-while-training-in-tensorflow-9fa1a1875a5

I was walked through how to build a class that inherits the Keras Callbacks class and is used to store the logs and metrics during training and how to create a callback that can use the matplotlib.pyplot to update a graph after each epoch. The callback is the users work in the link above, all else was done by me.

Thanks!